+++ test/type/Matrix.test.js
@@ -64,7 +64,7 @@ describe('matrix', function() {
     it('should resize the matrix correctly', function() {
       var m = new Matrix([[1,2,3],[4,5,6]]);
       m.resize([2,4]);
-      assert.deepEqual(m.valueOf(), [arr(1,2,3,uninit),arr(4,5,6,uninit)]);
+      assert.deepEqual(m.valueOf(), [[1,2,3,0], [4,5,6,0]]);
 
       m.resize([1,2]);
       assert.deepEqual(m.valueOf(), [[1,2]]);
@@ -90,6 +90,11 @@ describe('matrix', function() {
       assert.deepEqual(m.valueOf(), []);
     });
 
+    it('should resize the matrix with uninitialized default value', function() {
+      var m = new Matrix([]);
+      m.resize([3], math.uninitialized);
+      assert.deepEqual(m.valueOf(), arr(uninit, uninit, uninit));
+    });
   });
 
   describe('get', function() {
@@ -133,13 +138,13 @@ describe('matrix', function() {
       m.set([0, 2], 4);
       assert.deepEqual(m, new Matrix([
         [0, 0, 4],
-        arr(5, 0, uninit)
+        [5, 0, 0]
       ]));
 
       m.set([0,0,1], 3);
       assert.deepEqual(m, new Matrix([
-        [arr(0,3), arr(0,uninit), arr(4,uninit)],
-        [arr(5,uninit), arr(0,uninit), arr(undefined,uninit)]
+        [[0,3], [0,0], [4,0]],
+        [[5,0], [0,0], [0,0]]
       ]));
     });
 
@@ -242,19 +247,19 @@ describe('matrix', function() {
       m = new Matrix();
       m.resize([3,3]);
       assert.deepEqual(m, new Matrix([
-        arr(uninit, uninit, uninit),
-        arr(uninit, uninit, uninit),
-        arr(uninit, uninit, uninit)
+        [0, 0, 0],
+        [0, 0, 0],
+        [0, 0, 0]
       ]));
       m.subset(index([1,3], [1,3]), [[1,2],[3,4]]);
       assert.deepEqual(m, new Matrix([
-        arr(uninit, uninit, uninit),
-        arr(uninit, 1,2),
-        arr(uninit, 3,4)]));
+        [0, 0, 0],
+        [0, 1, 2],
+        [0, 3, 4]]));
       m.subset(index(0, [0,3]), [5,6,7]);
-      assert.deepEqual(m, new Matrix([[5,6,7],arr(uninit,1,2),arr(uninit,3,4)]));
+      assert.deepEqual(m, new Matrix([[5,6,7],[0,1,2],[0,3,4]]));
       m.subset(index([0,3], 0), [8,9,10]);  // unsqueezes the submatrix
-      assert.deepEqual(m, new Matrix([[8,6,7],arr(9,1,2),arr(10,3,4)]));
+      assert.deepEqual(m, new Matrix([[8,6,7],[9,1,2],[10,3,4]]));
     });
 
     it('should set the given subset with defaultValue for new elements', function() {
@@ -325,35 +330,35 @@ describe('matrix', function() {
       // set 2-dimensional with resize
       var m = new Matrix([[123]]);
       m.subset(index([1,3], [1,3]), [[1,2],[3,4]]);
-      assert.deepEqual(m, new Matrix([arr(123,uninit, uninit),arr(uninit,1,2),arr(uninit,3,4)]));
+      assert.deepEqual(m, new Matrix([[123,0,0],[0,1,2],[0,3,4]]));
 
       // set resize dimensions
       m = new Matrix([123]);
       assert.deepEqual(m.size(), [1]);
 
       m.subset(index([1,3], [1,3]), [[1,2],[3,4]]);
-      assert.deepEqual(m, new Matrix([arr(123,uninit,uninit),arr(uninit,1,2),arr(uninit,3,4)]));
+      assert.deepEqual(m, new Matrix([[123,0,0],[0,1,2],[0,3,4]]));
 
       m.subset(index([0,2], [0,2]), [[55,55],[55,55]]);
-      assert.deepEqual(m, new Matrix([arr(55,55,uninit),[55,55,2],arr(uninit,3,4)]));
+      assert.deepEqual(m, new Matrix([[55,55,0],[55,55,2],[0,3,4]]));
 
       m = new Matrix();
       m.subset(index([1,3], [1,3], [1,3]), [[[1,2],[3,4]],[[5,6],[7,8]]]);
       var res = new Matrix([
         [
-          arr(uninit, uninit, uninit),
-          arr(uninit, uninit, uninit),
-          arr(uninit, uninit, uninit)
+          [0, 0, 0],
+          [0, 0, 0],
+          [0, 0, 0]
         ],
         [
-          arr(uninit, uninit, uninit),
-          arr(uninit, 1, 2),
-          arr(uninit, 3, 4)
+          [0, 0, 0],
+          [0, 1, 2],
+          [0, 3, 4]
         ],
         [
-          arr(uninit, uninit, uninit),
-          arr(uninit, 5, 6),
-          arr(uninit, 7, 8)
+          [0, 0, 0],
+          [0, 5, 6],
+          [0, 7, 8]
         ]
       ]);
       assert.deepEqual(m, res);